MWP/requirements.txt
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128

# HuggingFace (CLIP Model/Tokenizer)
transformers==4.57.3
safetensors==0.7.0
tokenizers==0.22.1

# Data processing & analysis
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0
scipy==1.16.3

# Visualization
matplotlib==3.10.8
seaborn==0.13.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2

# Progress bar & utility
tqdm==4.67.1
Pillow==12.0.0